requirements.txt
Flask==3.0.3
google-analytics-data==0.18.2
gunicorn==22.0.0
